Refactoring remove product
