+++ src/app/services/database/product.service.ts
@@ -2,10 +2,8 @@ import { Injectable } from '@angular/core';
 import { Firestore, runTransaction } from '@angular/fire/firestore';
 import { ShoppingCartService } from 'src/app/services/database/shopping-cart.service';
 
-import { firstValueFrom, take } from 'rxjs';
 import { DbProduct } from './../../model/db-product';
 import { AbstractCrudService } from './abstract-crud.service';
-import { ShoppingCartProductService } from './shopping-cart-product.service';
 
 @Injectable({
   providedIn: 'root'
@@ -26,22 +24,11 @@ export class ProductService extends AbstractCrudService<DbProduct> {
    * @returns A promise that completes when the process has finished.
    */
   async removeProductFromAllCarts(productId: string) {
-    let shoppingCartIds = await this.shoppingCartService.getIds();
-    let productInfos = await Promise.all(shoppingCartIds.map(async shoppingCartId => {
-      const productIds = (await firstValueFrom(
-        (new ShoppingCartProductService(shoppingCartId, this.firestore)).getAll().pipe(take(1))
-      )).map(product => product.id);
-      return {
-        shoppingCartId: shoppingCartId,
-        productIds: productIds
-      };
-    }));
+    let shoppingCartAndProductIds = await this.shoppingCartService.getAllShoppingCartAndProductIds();
 
     await runTransaction(this.firestore, async (transaction) => {
       this.deleteT(transaction, productId);
-      productInfos.forEach(productInfo =>
-        this.shoppingCartService.removeProduct(transaction, productInfo.productIds, productId, productInfo.shoppingCartId)
-      );
+      this.shoppingCartService.removeProductFromAllCartsT(transaction, shoppingCartAndProductIds, productId);
     });
   }
